Play around with sinon
